Messages CRUD edited

+++ src/messages/messages.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable } from '@nestjs/common';
 import { CreateMessageDto } from './dto/create-message.dto';
 import { UpdateMessageDto } from './dto/update-message.dto';
 import { Repository } from 'typeorm';
@@ -20,57 +20,73 @@ export class MessagesService {
   }
 
   async findAll() {
-    const messages = await this.messagesRepository.find({
-      relations: {
-        author: true,
-        room: true,
-      },
-    });
+    const messages = await this.messagesRepository
+      .createQueryBuilder('messages')
+      .leftJoinAndSelect('messages.author', 'author')
+      .leftJoinAndSelect('messages.room', 'room')
+      .select([
+        'messages',
+        'author.id',
+        'author.fullName',
+        'room.id',
+        'room.title',
+      ])
+      .getMany();
 
-    return messages.map((message) => {
-      return {
-        ...message,
-        author: {
-          id: message.author.id,
-          fullName: message.author.fullName,
-        },
-        room: {
-          id: message.room.id,
-          title: message.room.title,
-        },
-      };
-    });
+    return messages;
   }
 
   async findOne(id: number) {
+    const message = await this.messagesRepository
+      .createQueryBuilder('message')
+      .where('message.id = :id', { id })
+      .leftJoinAndSelect('message.author', 'author')
+      .leftJoinAndSelect('message.room', 'room')
+      .select([
+        'message',
+        'author.id',
+        'author.fullName',
+        'room.id',
+        'room.title',
+      ])
+      .getOne();
+
+    return message;
+  }
+
+  async update(id: number, dto: UpdateMessageDto, userId: number) {
     const message = await this.messagesRepository.findOne({
       where: {
         id,
       },
       relations: {
         author: true,
-        room: true,
       },
     });
 
-    return {
-      ...message,
-      author: {
-        id: message.author.id,
-        fullName: message.author.fullName,
-      },
-      room: {
-        id: message.room.id,
-        title: message.room.title,
-      },
-    };
-  }
+    if (message.author.id !== userId) {
+      throw new ForbiddenException(
+        'У вас нет доступа к редактирования сообщения',
+      );
+    }
 
-  update(id: number, dto: UpdateMessageDto) {
     return this.messagesRepository.update(id, dto);
   }
 
-  remove(id: number) {
+  async remove(id: number, userId: number) {
+    const message = await this.messagesRepository.findOne({
+      where: {
+        id,
+      },
+      relations: {
+        author: true,
+      },
+    });
+
+    if (message.author.id !== userId) {
+      throw new ForbiddenException('У вас нет доступа к удалению сообщения');
+    }
+
     return this.messagesRepository.delete(id);
   }
 }
